requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
requests==2.31.0
openai==1.3.0
chromadb==0.4.18
pymupdf==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0
pdfplumber==0.10.2
pypdf==3.17.1
asyncio==3.4.3
uvicorn
python-multipart
pytesseract
pydantic
PyMuPDF
